requirements.txt
numpy>=1.24
PyYAML>=6.0
gymnasium>=0.29
torch>=2.1
matplotlib>=3.7
pandas>=2.0
pytest>=7.0
